Green - return 50 to pass

+++ Checkout.java
@@ -3,7 +3,7 @@ package cleancode;
 public class Checkout {
 
     public Integer calculateTotalPrice(String itemCode) {
-        return new Integer(0);
+        return new Integer(50);
     }
 
 }
